test: fix pylint errors
Fix pylint errors in test files:
* move fixtures to conftest.py to avoid redefinition
* add missing module and function docstrings
* fix order of imports
* drop unused imports
* fix invalid name error
